[#7971] docs, review suggestions
