+++ ckanext/datastore/interfaces.py
@@ -207,13 +207,13 @@ class IDataDictionaryForm(interfaces.Interface):
         data for that plugin.
 
         e.g. a statistics plugin that needs to store per-column information
-        might store this with plugin_data by inserting values like:
+        might store this with plugin_data by inserting values like::
 
-            {0: {'statistics': {'minimum': 34, ...}, ...}, ...}
+          {0: {'statistics': {'minimum': 34, ...}, ...}, ...}
 
-                                ^ the data stored for this field+plugin
-                  ^ the name of the plugin
-             ^ 0 for the first field passed in fields
+          #                   ^ the data stored for this field+plugin
+          #     ^ the name of the plugin
+          #^ 0 for the first field passed in fields
 
         Values not removed from field info by validation will be available in
         the field `info` dict returned from `datastore_search` and
