+++ doc/extensions/index.rst
@@ -41,3 +41,4 @@ features by developing your own CKAN extensions.
    translating-extensions
    flask-migration
    signals
+   custom-data-dictionary
